refactor 1.5.2: Input argument type should be unsigned int. Return type should be long long.

+++ Exercises/Level1/Section_1.5/Exercise_2/ExerciseTwo.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 
-long factorial(int n) {
+long long factorial(unsigned int n) {
     if (n <= 1) return 1; // Base case
-    return n * factorial(n - 1); // Recursive call
+    return (long long)n * factorial(n - 1); // Recursive call
 }
 
 int main() {
-    int number = 6;
+    unsigned int number = 6;
     printf("Factorial of %d is %ld\n", number, factorial(number));
     return 0;
 }
